Fix chunk boundary issue with barrels
Barrels which had their block in on chunk and their itemframe in other could get
taken for corrupted/missing barrels due to the itemframe being in an unloaded
chunk. This change fixes that problem by detecting when that could happen and
deferring the finding of the itemframe until later.

The facing direction and itemsstack are now also saved in the barrel state
to be more robust in the future.

Barrel conduit and barrel controllers also used to trugger loading of chunks
as they search adjacent blocks, this has been fixed.